Flag match requirement misspellings early on
Also added a unit test about such misspelled match requirements.

Misspelling a match requirement, like for example, specifying `header`
instead of `headers` would make the test silently succeed when it should
maybe have failed.

Ensure we raise an exception in such cases, to flag that the match
requirement is incorrect, so this category of errors is caught early on.